almost done with signup and login. only need to implement session creations.
Did a "major" refactor of code so that when an
instance of User() is created, user_id, username,
and email, are guranteed to be present.

When a user signs up, an instance of New_User() is
created and then immediately saved to the database.
Then you can create User() object to have access to
the user, along with the guranteed basic info